requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
PyMySQL==1.1.0
Werkzeug==2.3.6
cryptography==41.0.3